sets up conftest.py for fixture sharing
